+++ src/js/components/all-read.test.js
@@ -1,7 +1,7 @@
 import React from 'react'; // eslint-disable-line no-unused-vars
 import { mount } from 'enzyme';
 
-import AllRead from '../../components/all-read';
+import AllRead from './all-read';
 
 describe('components/all-read.js', function() {
   it('should render itself & its children', function() {
