+++ src/js/components/notifications.test.js
@@ -2,10 +2,10 @@ import React from 'react'; // eslint-disable-line no-unused-vars
 import renderer from 'react-test-renderer';
 import { List } from 'immutable';
 
-import AccountNotifications from '../../components/notifications';
-import { mockedGithubNotifications } from '../../__mocks__/mockedData';
+import AccountNotifications from './notifications';
+import { mockedGithubNotifications } from '../__mocks__/mockedData';
 
-jest.mock('../../components/repository');
+jest.mock('./repository');
 
 describe('components/notifications.js', () => {
   it('should render itself (github.com with notifications)', () => {
